added Cmake and clang-{format,tidy}

+++ str_utils.c
@@ -5,14 +5,14 @@ bool next_token(Spliter *liner, str_view *line) {
     return false;
   }
   char *start = liner->pos;
-  char *end = strchr(start, liner->delim);
+  char *end   = strchr(start, liner->delim);
   if (end == NULL || (end >= (liner->fc.data + liner->fc.len))) {
     line->data = start;
-    line->len = liner->fc.len - (liner->pos - liner->fc.data);
+    line->len  = liner->fc.len - (liner->pos - liner->fc.data);
     liner->pos = liner->fc.data + liner->fc.len;
   } else {
     line->data = start;
-    line->len = end - start;
+    line->len  = end - start;
     liner->pos = end + 1;
   }
 
@@ -20,7 +20,7 @@ bool next_token(Spliter *liner, str_view *line) {
 }
 
 int str_find_char(const str_view *sv, predicate pred) {
-  for (int i = 0; i < sv->len; i++) {
+  for (size_t i = 0; i < sv->len; i++) {
     if (pred(sv->data[i])) {
       return i;
     }
@@ -29,7 +29,7 @@ int str_find_char(const str_view *sv, predicate pred) {
 }
 
 int str_find_last_char(const str_view *sv, predicate pred) {
-  for (int i = sv->len - 1; i >= 0; i--) {
+  for (size_t i = sv->len - 1; i >= 0; i--) {
     if (pred(sv->data[i])) {
       return i;
     }
